Add files via upload
This repository showcases a series of experiments in computer vision and machine learning. The primary focus was on developing a real-time face recognition system using OpenCV. The project involved collecting a dataset of facial images, training a deep learning model, and implementing a user interface for live video analysis. Additionally, exploratory data analysis was conducted on various datasets to gain insights into data distributions and relationships.